# packages/contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/query-plans/{planId}/execute": {
      "post": {
        "operationId": "DiscoveryController_execute_v1",
        "parameters": [
          {
            "name": "planId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "runId",
                        "status"
                      ],
                      "properties": {
                        "runId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "执行 READY 查询计划：多源发现 → Raw → Canonical（异步，Temporal 编排）",
        "tags": [
          "Discovery"
        ],
        "x-required-scopes": [
          "acquisition:write"
        ]
      }
    },
    "/api/v1/discovery-runs/{runId}": {
      "get": {
        "operationId": "DiscoveryController_getRun_v1",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "id",
                        "planId",
                        "icpId",
                        "status",
                        "stats",
                        "createdAt",
                        "completedAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "planId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "icpId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "status": {
                          "type": "string"
                        },
                        "stats": {
                          "type": "object",
                          "additionalProperties": true
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "completedAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "发现执行状态与统计（每源计数/归一/Suppression）",
        "tags": [
          "Discovery"
        ],
        "x-required-scopes": [
          "acquisition:read"
        ]
      }
    },
    "/api/v1/canonical-companies": {
      "get": {
        "operationId": "DiscoveryController_listCompanies_v1",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "page"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "description": "CanonicalCompany（归一视图 + 联系人 + 字段级 Evidence）"
                      }
                    },
                    "page": {
                      "type": "object",
                      "required": [
                        "next_cursor",
                        "has_more"
                      ],
                      "properties": {
                        "next_cursor": {
                          "type": "string",
                          "nullable": true,
                          "description": "下一页游标；null = 没有更多"
                        },
                        "has_more": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "发现的目标客户公司（归一后，游标分页；?status=NEW|ENRICHED|SUPPRESSED）",
        "tags": [
          "Discovery"
        ],
        "x-required-scopes": [
          "acquisition:read"
        ]
      }
    },
    "/api/v1/canonical-companies/{id}": {
      "get": {
        "operationId": "DiscoveryController_getCompany_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "CanonicalCompany（归一视图 + 联系人 + 字段级 Evidence）"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "公司详情：canonical 视图 + 联系人 + 字段级 Evidence（每个字段值的来源）",
        "tags": [
          "Discovery"
        ],
        "x-required-scopes": [
          "acquisition:read",
          "personal-data:read"
        ]
      }
    },
    "/api/v1/canonical-companies/{id}/discover-contacts": {
      "post": {
        "operationId": "DiscoveryController_discoverContacts_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "联系人发现结果（新建联系人/联系点计数）"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "按需发现联系人（Waterfall 第5步：仅高价值企业；Suppression 先行过滤）",
        "tags": [
          "Discovery"
        ],
        "x-required-scopes": [
          "acquisition:write",
          "personal-data:read",
          "compliance:manage"
        ]
      }
    },
    "/api/v1/contact-points/{pointId}/verify": {
      "post": {
        "description": "合规门：职能邮箱默认自动验证；人名邮箱（个人数据）需 lawfulBasis 或 allowPersonalWithoutBasis，否则 BLOCKED（不探测）。",
        "operationId": "DiscoveryController_verify_v1",
        "parameters": [
          {
            "name": "pointId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyContactPointDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "验证结果（status + 探测细节留痕）"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "邮箱验证（Waterfall 第7步）：状态回写 UNVERIFIED→VALID|RISKY|INVALID|BLOCKED",
        "tags": [
          "Discovery"
        ],
        "x-required-scopes": [
          "acquisition:write",
          "personal-data:read",
          "compliance:manage"
        ]
      }
    },
    "/api/v1/canonical-companies/{id}/guess-emails": {
      "post": {
        "description": "合规门：猜出的都是人名邮箱（个人数据），需 lawfulBasis 或 allowPersonalWithoutBasis，否则一律 blocked（零探测）。RISKY 未证实猜测落库但 allowedActions 不含 outreach（不可群发）；suppression 命中不落。",
        "operationId": "DiscoveryController_guessEmails_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuessEmailsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "邮箱猜测结果（补全计数 + 逐决策人状态/落库态）"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "猜测缺邮箱决策人的邮箱（排列/格式学习 + SMTP RCPT 验证 → 落库）",
        "tags": [
          "Discovery"
        ],
        "x-required-scopes": [
          "acquisition:write",
          "personal-data:read",
          "compliance:manage"
        ]
      }
    },
    "/api/v1/suppressions": {
      "post": {
        "operationId": "DiscoveryController_addSuppression_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSuppressionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Suppression 记录"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "加入禁联名单（email/domain/company_name）；命中的公司立即 SUPPRESSED",
        "tags": [
          "Discovery"
        ],
        "x-required-scopes": [
          "compliance:manage"
        ]
      },
      "get": {
        "operationId": "DiscoveryController_listSuppressions_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "description": "Suppression 记录"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "禁联名单",
        "tags": [
          "Discovery"
        ],
        "x-required-scopes": [
          "compliance:manage"
        ]
      }
    },
    "/api/v1/suppressions/{id}": {
      "delete": {
        "operationId": "DiscoveryController_removeSuppression_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "deleted"
                      ],
                      "properties": {
                        "deleted": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "移除禁联记录",
        "tags": [
          "Discovery"
        ],
        "x-required-scopes": [
          "compliance:manage"
        ]
      }
    },
    "/api/v1/data-providers": {
      "get": {
        "operationId": "DiscoveryController_listProviders_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "description": "DataProvider（源/状态/成本）"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Provider 注册表（平台级：状态/成本；DISABLED = Kill Switch）",
        "tags": [
          "Discovery"
        ],
        "x-required-scopes": [
          "ops:read"
        ]
      }
    },
...
  },
  "info": {
    "title": "Global API",
    "description": "出海企业 AI 全球客户开发与增长平台 · 后端 API（前端接入见 packages/contracts/INTEGRATION.md）",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "同源部署（相对路径；具体 host 由部署环境决定）"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "VerifyContactPointDto": {
        "type": "object",
        "properties": {
          "lawfulBasis": {
            "type": "string",
            "enum": [
              "legitimate_interest",
              "consent",
              "contract",
              "legal_obligation"
            ],
            "description": "探测人名邮箱的合法性基础（GDPR Art.6）；职能邮箱可省略"
          },
          "lawfulBasisRef": {
            "type": "string",
            "description": "LIA / 工单 / 合同 / 同意记录的引用（可审计）"
          },
          "lawfulBasisNote": {
            "type": "string",
            "description": "备注"
          },
          "allowPersonalWithoutBasis": {
            "type": "boolean",
            "description": "显式开关：无 lawfulBasis 也允许探测人名邮箱（默认 false，仍留痕）"
          }
        }
      },
      "GuessEmailsDto": {
        "type": "object",
        "properties": {
          "lawfulBasis": {
            "type": "string",
            "enum": [
              "legitimate_interest",
              "consent",
              "contract",
              "legal_obligation"
            ],
            "description": "探测人名邮箱的合法性基础（GDPR Art.6）；猜出的都是人名邮箱"
          },
          "lawfulBasisRef": {
            "type": "string",
            "description": "LIA / 工单 / 合同 / 同意记录的引用（可审计）"
          },
          "lawfulBasisNote": {
            "type": "string",
            "description": "备注"
          },
          "allowPersonalWithoutBasis": {
            "type": "boolean",
            "description": "显式开关：无 lawfulBasis 也允许探测人名邮箱（默认 false，仍留痕）"
          },
          "maxContacts": {
            "type": "number",
            "description": "最多补全几个缺邮箱决策人（有界护栏，默认 25）"
          },
          "maxProbe": {
            "type": "number",
            "description": "每人最多探测几个候选（有界护栏，默认 8）"
          }
        }
      },
      "CreateSuppressionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "email",
              "domain",
              "company_name"
            ]
          },
          "value": {
            "type": "string",
            "example": "noreply@example.com"
          },
          "reason": {
            "type": "string",
            "enum": [
              "unsubscribe",
              "bounce",
              "complaint",
              "manual",
              "legal"
            ]
          }
        },
        "required": [
          "type",
          "value"
        ]
      },
...
    }
  }
}